integer type id columns changed as long type

+++ src/main/java/com/house/transport/repository/MoverRepository.java
@@ -3,5 +3,5 @@ package com.house.transport.repository;
 import com.house.transport.model.Mover;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MoverRepository extends JpaRepository<Mover, Integer> {
+public interface MoverRepository extends JpaRepository<Mover, Long> {
 }
